+++ lib/mustache.mli
@@ -22,10 +22,11 @@ type dotted_name = string list
 type t =
   | String of string
   | Escaped of dotted_name
-  | Section of section
   | Unescaped of dotted_name
-  | Partial of partial
+  | Section of section
   | Inverted_section of section
+  | Partial of partial
+  | Param of param
   | Concat of t list
   | Comment of string
 and section =
@@ -34,7 +35,12 @@ and section =
 and partial =
   { indent: int;
     name: name;
+    params: param list option;
     contents: t option Lazy.t }
+and param =
+  { indent: int;
+    name: name;
+    contents: t }
 
 type loc =
     { loc_start: Lexing.position;
@@ -129,13 +135,15 @@ val render :
     @param string Applied to each literal part of the template.
     @param escaped Applied to ["name"] for occurrences of [{{name}}].
     @param unescaped Applied to ["name"] for occurrences of [{{{name}}}].
-    @param partial Applied to ["box"] for occurrences of [{{> box}}].
+    @param partial Applied to ["box"] for occurrences of [{{> box}}] or [{{< box}}].
+    @param params Applied to ["param"] for occurrences of [{{$ param}}].
     @param comment Applied to ["comment"] for occurrences of [{{! comment}}]. *)
 val fold : string: (string -> 'a) ->
   section: (inverted:bool -> dotted_name -> 'a -> 'a) ->
   escaped: (dotted_name -> 'a) ->
   unescaped: (dotted_name -> 'a) ->
-  partial: (int -> name -> t option Lazy.t -> 'a) ->
+  partial: (?indent:int -> name -> ?params:(int * name * 'a) list -> t option Lazy.t -> 'a) ->
+  param: (?indent:int -> name -> 'a -> 'a) ->
   comment: (string -> 'a) ->
   concat:('a list -> 'a) ->
   t -> 'a
@@ -170,8 +178,20 @@ val inverted_section : dotted_name -> t -> t
 (** [{{#person}} {{/person}}] *)
 val section : dotted_name -> t -> t
 
-(** [{{> box}}] *)
-val partial : ?indent:int -> name -> t option Lazy.t -> t
+(** [{{> box}}]
+    or
+    {[
+     {{< box}}
+       {{$param1}} default value for param1 {{/param1}}
+       {{$param2}} default value for param1 {{/param2}}
+     {{/box}}
+    ]}
+ *)
+val partial :
+  ?indent:int -> name -> ?params:(int * name * t) list -> t option Lazy.t -> t
+
+(** [{{$foo}} {{/foo}}] *)
+val param : ?indent:int -> name -> t -> t
 
 (** [{{! this is a comment}}] *)
 val comment : string -> t
@@ -189,10 +209,11 @@ module With_locations : sig
   type desc =
     | String of string
     | Escaped of dotted_name
-    | Section of section
     | Unescaped of dotted_name
-    | Partial of partial
+    | Section of section
     | Inverted_section of section
+    | Partial of partial
+    | Param of param
     | Concat of t list
     | Comment of string
   and section =
@@ -201,7 +222,12 @@ module With_locations : sig
   and partial =
     { indent: int;
       name: name;
+      params: param list option;
       contents: t option Lazy.t }
+  and param =
+    { indent: int;
+      name: name;
+      contents: t }
   and t =
     { loc : loc;
       desc : desc }
@@ -266,13 +292,15 @@ module With_locations : sig
       @param string Applied to each literal part of the template.
       @param escaped Applied to ["name"] for occurrences of [{{name}}].
       @param unescaped Applied to ["name"] for occurrences of [{{{name}}}].
-      @param partial Applied to ["box"] for occurrences of [{{> box}}].
+      @param partial Applied to ["box"] for occurrences of [{{> box}}] or [{{< box}}].
+      @param params Applied to ["param"] for occurrences of [{{$ param}}].
       @param comment Applied to ["comment"] for occurrences of [{{! comment}}]. *)
   val fold : string: (loc:loc -> string -> 'a) ->
     section: (loc:loc -> inverted:bool -> dotted_name -> 'a -> 'a) ->
     escaped: (loc:loc -> dotted_name -> 'a) ->
     unescaped: (loc:loc -> dotted_name -> 'a) ->
-    partial: (loc:loc -> int -> name -> t option Lazy.t -> 'a) ->
+    partial: (loc:loc -> ?indent:int -> name -> ?params:(int * name * 'a) list -> t option Lazy.t -> 'a) ->
+    param: (loc:loc -> ?indent:int -> name -> 'a -> 'a) ->
     comment: (loc:loc -> string -> 'a) ->
     concat:(loc:loc -> 'a list -> 'a) ->
     t -> 'a
@@ -305,8 +333,20 @@ module With_locations : sig
   (** [{{#person}} {{/person}}] *)
   val section : loc:loc -> dotted_name -> t -> t
 
-  (** [{{> box}}] *)
-  val partial : loc:loc -> ?indent:int -> name -> t option Lazy.t -> t
+  (** [{{> box}}]
+      or
+      {[
+       {{< box}}
+         {{$param1}} default value for param1 {{/param1}}
+         {{$param2}} default value for param1 {{/param2}}
+       {{/box}}
+      ]}
+   *)
+  val partial :
+    loc:loc -> ?indent:int -> name -> ?params:(int * name * t) list -> t option Lazy.t -> t
+
+  (** [{{$foo}} {{/foo}}] *)
+  val param : loc:loc -> ?indent:int -> name -> t -> t
 
   (** [{{! this is a comment}}] *)
   val comment : loc:loc -> string -> t